Remove ScalarArrayDataR, instead implement Storable for Vec
Arrays of Vecs were previously typed as Arrays of the elements of the Vec. This caused that many functions needed a separate case for Vecs.

With this change, that is no longer the case. An array of Vecs is just an array of Vecs (on the type level). Note that this commit doesn't change the in-memory representation of arrays.

Uniformity between Vector types and Single types is introduced by implementing Storable for Vec.