Fix output size of foldSeg
The desugaring of foldSeg allocated an output array with the same size as the input, which was not correct. The output size should be the number of segments. This commit fixes that.

Note that this function gets an array of offsets of the segments, and that array is one more than the number of segments. We thus need to subtract by one, and we might in the future want to write a compiler pass that (in OperationAcc) simplifies:

let y = compute x + 1
let z = compute y - 1

to

let y = compute x + 1
let z = compute x