Also detect invariant variables in the loop by comparing with initial value
Previously, a variable in the state of a while loop was considered as invariant if the step function returned that variable unchanged. Now we check if it returns it unchanged *or returns the initial value*. The latter occurs in some patterns for the size variables of arrays.

This analysis then allows us to remove these variables from the state of the while loop (which we already did before this commit)