Let 'expand' fuse better, and use strict && and ||
This commit makes some changes to the definition of 'expand', and the expression simplifier, to let the compiler better optimize 'expand'. It can now deduce that some arrays have the same size, which then allows fusion to do its work better.

Also fixes alignment and changes 'and' and 'or' to use strict versions of && and ||.